requirements.txt
Click==7.0
numpy==1.17.3
Pillow==6.2.1
pytorch-ignite==0.2.1
ruamel.yaml==0.16.5
torch==1.3.1
tensorboard==2.0.2
tensorboardX==1.9
torchvision==0.4.2
tqdm==4.40.1